Rog and Wiz Fixes
Rog was passing Vision discSpell as UserData. Added a variable to convert to string on line 393
Wiz missing DoManaBurn. Added to self.settings. 

Note: Wiz spell gems 1 and 2 are returning "ComboClaw, ComboEtherealNuke, and ComboNuke" when set to Combo mode. Returning error that these don't exist in AbilitySet. Not sure what the intended functionality of Combo mode is.