Berserker snare ability active
This has the Berserker do the Leg Strike ability on the rotation priority list (low down the list). It will handle summoning the Blunt Axes necessary for this ability.

Note: This is not ready for prime time as it should dynamically use the better snare ability (Leg Cut) at appropriate level and add a toggle setting so that snare can be turned on/off.