added option to import a complete custom config
if you comment out sections and try to import the config as a custom one, the commented out sections get ignored.

Added a flag to the config FullConfig = true/false
if the flag is present and true then replace the default config with the custom one. otherwise use the mergeTables function.

local _ClassConfig = {
    _version          = "1.0 Beta",
    _author           = "Derple",
    ['FullConfig'] = true,