MyPaths Path Support
If a MyPaths Paths file is found Load it into a table.

if the current zone has any paths from MyPaths display a drop down list to select from, and populate the waypoints table with the points from the selected path.

also has a button to refresh the Paths file incase you made a new one or changed it in MyPaths