Support for MyPaths Paths in Farm Mode
If the MyPaths_Paths.lua is present then load a table for it.

if the zone has any paths then show a drop down to select from or use the original saved waypoints.

you can reload the paths file with a button click incase it any were edited or imported / added in MyPaths.