Auto stash before checking out "origin/main"
if you had buffing turned off but DoPet on the pet wasn't being summoned because we  never entered those rotations.

added one just to manage the pet summoning when you don't have one and DoPet is on.

can add other things to this that you would  like to handle under these circumstances as well