removed autoloot setting
cleaned up some ini stuff
added back in the save to ini functions. for when looting new items or auto sell/tribute
also writes to the ini if we add items to global or buy

Does not write to the file when we modify or delete an item. that stays in our local rgmers settings we  use. this way we do not pollute the loot.ini