Loot Updates
Added Pop up window option for new items.
Added Links to the loot tables. (these will populate for items missing them as you loot items)
Added Toggle to only loot a corpse once.

The new items pop up allows you to set items rules yourself for items not already in the table. The fields will be prefilled with your evaluated settings.

If the item was not no drop then we will make the decision and loot accordingly, and show the window.

If you have only loot corpse on and a new item appears, saving the settings from the new item popup will unflag that corpse incase there was a nodrop item on the corpse and we wanted to loot it.

Window includes  the following information:
* Link button,
* Item Name,
* Nodrop,
* Lore,
* Augment,
* TS,
* Sell Price.