bulk edit, toggle INFO messages off, Actors work
You can now Bulk set rules for entire pages of items

refined the number of items to show down to increments of 10 to help.

you can now toggle the INFO messages off.
these also now go to the lootnscoot console and not the main mq console

actors should properly communicate with standalone and mercs clients on the same machine.