pull check for swarm, aa healer, and doppelganger  pets
in the pull deny list check for swarm pets, doppelgangers, and healers aa healing pets.

the typical check fails on these spawns but we can check the surname for 's Pet etc.

this should stop the tank for getting hung up trying to pull SomeCleric`s Pet  Benediction pet as well as enchanter doppelgangers and swarm pets.

these pets also use `s in the cleanname and 's in the surname for some odd reason.