Mez Update (#701)
Refactored Mez code to wait for a reasonable time period if we are currently casting, moving, or in the global cooldown.
ENC: while AE mezzing, Beam of Slumber should only fire if we are in the global cooldown or if the AE mez spell gem timer is not ready.
BRD: if we are ST mezzing, we will always attempt to use Dirge of the Sleepwalker before we waste time on a song.
Allowed AE Mez to continue when the target dies.
Reduced retry attempts on all mezzes.
Ensured we turned attack off when updating the mez list.
Added another failsafe to switch to AE mez while under certain conditions while we are waiting for ST mez to be ready.
Added an additional autotarget ID check before we perform a single target mez.
Known Issue/Limitation: During mez processing, we do not update our autotarget; during situations of rapid autotarget change, there will be situations in which we attempt to use a ST mez on what may be everyone else's autotarget. No adjustment is currently planned to address this behavior.
Adjusted log messages.