Added support for buffing your group with items
On THF, there are a lot of clickies that replace buff spells, like DI, etc.

I added GroupBuffItemCheck to handle this, which is no problem; it follows the same rules as the rest of the modular buff check functions.

My issue was that nothing worked.

After chasing things down, it seems what was happening was the item entry name was being thrown into mq.TLO.Spell(entry.name) and returning null, which was killing the groupbuffcheck.

I've made some adjustments to condition args checks that I would like a second eye on to alleviate this. If you look at the old code, you can see where an item used outside of a mapping would trigger what I described.

Everything appears to function normally, I added discs to the list of things that returned a spell type as well, I'm not sure why they would be omitted.

The other casting utils changes are just something I noticed along the way.